#23 Add equation creation to api

+++ SpaceY/Interface/RestEquationParam.cs
@@ -18,6 +18,6 @@
         /// <summary>
         /// Gets or sets the default value of a parameter.
         /// </summary>
-        public decimal Default { get; set; }
+        public double Default { get; set; }
     }
 }
